requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
mysql-connector-python==8.0.33
passlib==1.7.4
python-multipart==0.0.6